tweaks to testing suite
